WIP - refactoring common code
