+++ src/components/RepoValidationCard.js
@@ -3,8 +3,6 @@ import TreeView from '@material-ui/lab/TreeView'
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore'
 import ChevronRightIcon from '@material-ui/icons/ChevronRight'
 import TreeItem from '@material-ui/lab/TreeItem'
-import { Tooltip } from '@material-ui/core'
-import { IconButton } from '@material-ui/core'
 
 import PropTypes from 'prop-types'
 import { Card } from 'translation-helps-rcl'
@@ -18,15 +16,9 @@ import { checkTwForBook, checkTaForBook } from '@utils/checkArticles'
 import { WORKING, OK, REPO_NOT_FOUND, FILE_NOT_FOUND, BOOK_NOT_IN_MANIFEST, NO_TWL_REPO, SEE_TWL_ERROR, NO_TN_REPO, SEE_TN_ERROR } 
 from '@common/constants'
 
-import CreateIcon from '@material-ui/icons/Create'
-import DoneIcon from '@material-ui/icons/Done'
-import VisibilityIcon from '@material-ui/icons/Visibility'
-import BlockIcon from '@material-ui/icons/Block'
-
-import CreateRepoButton from './CreateRepoButton'
-import AddBookToManifest from './AddBookToManifest'
 import DenseTable from './DenseTable'
-import ViewListButton from './ViewListButton'
+import { checkManifestBook } from '@common/manifests'
+import { applyIcon } from './iconHelper'
 
 export default function RepoValidationCard({
   bookId,
@@ -112,44 +104,6 @@ export default function RepoValidationCard({
       setRefresh,
     }
   } = useContext(AdminContext)
-
-  function checkManifestBook(manifest, repoTree, setError) {
-    let projects = []
-    if (manifest && manifest.projects) {
-      projects = manifest.projects
-    } else {
-      return
-    }
-    let isBookIdInManfest = false
-    let pathToBook;
-    for (let i=0; i < projects.length; i++) {
-      if ( projects[i]?.identifier === bookId ) {
-        isBookIdInManfest = true
-        pathToBook = projects[i].path
-        break
-      }
-    }
-
-    // if project id exists, then does the file actually exist?
-    if ( isBookIdInManfest ) {
-      let _fileExists = false
-      for (let i=0; i < repoTree.length; i++) {
-        let _path = repoTree[i].path
-        let _manifestpath = pathToBook.replace(/^\.\//,'')
-        if ( _manifestpath === _path ) {
-          _fileExists = true
-          break
-        }
-      }
-      if ( _fileExists ) {
-        setError(OK)
-      } else {
-        setError(FILE_NOT_FOUND)
-      }
-    } else {
-      setError(BOOK_NOT_IN_MANIFEST)
-    }
-  }
 
   /*
   -- 
@@ -248,8 +202,8 @@ export default function RepoValidationCard({
   -- 
   */
   useEffect(() => {
-    checkManifestBook(tnRepoTreeManifest, tnRepoTree, setTnBookErrorMsg)
-  }, [tnRepoTree, tnRepoTreeManifest])
+    checkManifestBook(bookId, tnRepoTreeManifest, tnRepoTree, setTnBookErrorMsg)
+  }, [bookId, tnRepoTree, tnRepoTreeManifest])
 
   /*
   -- 
@@ -257,8 +211,8 @@ export default function RepoValidationCard({
   -- 
   */
   useEffect(() => {
-    checkManifestBook(twlRepoTreeManifest, twlRepoTree, setTwlBookErrorMsg)
-  }, [twlRepoTree, twlRepoTreeManifest])
+    checkManifestBook(bookId, twlRepoTreeManifest, twlRepoTree, setTwlBookErrorMsg)
+  }, [bookId, twlRepoTree, twlRepoTreeManifest])
 
   /*
   -- 
@@ -266,8 +220,8 @@ export default function RepoValidationCard({
   -- 
   */
   useEffect(() => {
-    checkManifestBook(ltRepoTreeManifest, ltRepoTree, setLtBookErrorMsg)
-  }, [ltRepoTree, ltRepoTreeManifest])
+    checkManifestBook(bookId, ltRepoTreeManifest, ltRepoTree, setLtBookErrorMsg)
+  }, [bookId, ltRepoTree, ltRepoTreeManifest])
 
   /*
   -- 
@@ -275,8 +229,8 @@ export default function RepoValidationCard({
   -- 
   */
   useEffect(() => {
-    checkManifestBook(stRepoTreeManifest, stRepoTree, setStBookErrorMsg)
-  }, [stRepoTree, stRepoTreeManifest])
+    checkManifestBook(bookId, stRepoTreeManifest, stRepoTree, setStBookErrorMsg)
+  }, [bookId, stRepoTree, stRepoTreeManifest])
 
   /*
   -- 
@@ -284,8 +238,8 @@ export default function RepoValidationCard({
   -- 
   */
   useEffect(() => {
-    checkManifestBook(tqRepoTreeManifest, tqRepoTree, setTqBookErrorMsg)
-  }, [tqRepoTree, tqRepoTreeManifest])
+    checkManifestBook(bookId, tqRepoTreeManifest, tqRepoTree, setTqBookErrorMsg)
+  }, [bookId, tqRepoTree, tqRepoTreeManifest])
 
   /*
   -- 
@@ -293,8 +247,8 @@ export default function RepoValidationCard({
   -- 
   */
   useEffect(() => {
-    checkManifestBook(sqRepoTreeManifest, sqRepoTree, setSqBookErrorMsg)
-  }, [sqRepoTree, sqRepoTreeManifest])
+    checkManifestBook(bookId, sqRepoTreeManifest, sqRepoTree, setSqBookErrorMsg)
+  }, [bookId, sqRepoTree, sqRepoTreeManifest])
 
   /*
   -- 
@@ -302,8 +256,8 @@ export default function RepoValidationCard({
   -- 
   */
   useEffect(() => {
-    checkManifestBook(snRepoTreeManifest, snRepoTree, setSnBookErrorMsg)
-  }, [snRepoTree, snRepoTreeManifest])
+    checkManifestBook(bookId, snRepoTreeManifest, snRepoTree, setSnBookErrorMsg)
+  }, [bookId, snRepoTree, snRepoTreeManifest])
 
   let _ltRepo = languageId
   let _stRepo = languageId
@@ -313,99 +267,6 @@ export default function RepoValidationCard({
   } else {
     _ltRepo += "_glt"
     _stRepo += "_gst"
-  }
-  const applyIcon = (repo,repoErr,bookErr,manifest,manifestSha) => {
-    // console.log("applyIcon() parameters:",`repo:${repo}
-    //   repoErr:${repoErr}
-    //   bookErr:${bookErr}
-    //   manifest:${manifest}
-    //   manifestSha:${manifestSha}
-    // `)
-    if ( repo.endsWith("_tw") || repo.endsWith("_ta") ) {
-      if ( repoErr === null && bookErr === WORKING ) {
-        return (
-          <p>{WORKING}</p>
-        )
-      }
-    }
-
-    if ( repoErr === null && bookErr === null ) {
-      return (
-        <p>{WORKING}</p>
-      )
-    }
-
-    if ( repoErr === REPO_NOT_FOUND ) {
-      return (
-        <CreateRepoButton active={true} server={server} owner={owner} repo={repo} refresh={refresh} bookId={bookId} onRefresh={setRefresh} />
-      )
-    }
-
-    if ( repoErr !== null ) {
-      return (
-        <p>{repoErr}</p>
-      )
-    }
-
-    if ( bookErr === OK ) {
-      return (
-        <DoneIcon />
-      )
-    }
-
-    if ( bookErr === BOOK_NOT_IN_MANIFEST ) {
-      return (
-        <AddBookToManifest 
-          active={true} server={server} owner={owner} repo={repo} refresh={refresh} 
-          manifest={manifest} sha={manifestSha} bookId={bookId} onRefresh={setRefresh} 
-        />
-      )
-    }
-
-    if ( bookErr.endsWith('Missing') ) {
-      if ( repo.endsWith("_tw") ) {
-        const title = "Translation Word Articles Missing"
-        return (
-          <ViewListButton title={title} value={twMissing} />
-        )
-      } else {
-        const title = "Translation Academy Articles Missing"
-        return (
-          <ViewListButton title={title} value={taMissing} />
-        )
-      }
-
-    }
-
-    if ( bookErr !== null ) {
-      if ( repo.endsWith("_tw") ) {
-        return (
-          <Tooltip title="Use tC Create to create translation word list">
-            <IconButton className={classes.iconButton} aria-label="Use-tc-create-tw">
-              <BlockIcon />
-            </IconButton>
-          </Tooltip>
-        )
-      }
-      if ( repo.endsWith("_ta") ) {
-        return (
-          <Tooltip title="Use tC Create to create translation notes">
-            <IconButton className={classes.iconButton} aria-label="Use-tc-create-ta">
-              <BlockIcon />
-            </IconButton>
-          </Tooltip>
-        )
-      }
-      return (
-        <Tooltip title="Use tC Create to create file">
-          <IconButton className={classes.iconButton} aria-label="Use-tc-create">
-            <CreateIcon/>
-          </IconButton>
-        </Tooltip>
-
-      )
-    }
-
   }
   const headers = ["Resource", "Repo", "Status", "Action"]
   const rows = [
