+++ src/hooks/useLtRepoValidation.js
@@ -10,7 +10,6 @@ export default function useLtRepoValidation({authentication, owner, server, lang
     setValues
   ] = useState({ltRepoTree:null, ltRepoTreeManifest:null, ltManifestSha:null, ltRepoTreeStatus:WAITING})
 
-  console.log("useLtRepoValidation(), refresh=",refresh)
   // Translation Notes Hook
   // Example: https://qa.door43.org/api/v1/repos/vi_gl/vi_lt/git/trees/master?recursive=true&per_page=99999
   useEffect(() => {
@@ -26,10 +25,8 @@ export default function useLtRepoValidation({authentication, owner, server, lang
       const {RepoTree: _tree, Manifest: _manifest, ManifestSha: _manifestSha, RepoTreeStatus: _errorMesage} =  await getTreesManifest(authentication, url)
       setValues({ltRepoTree: _tree, ltRepoTreeManifest: _manifest, ltManifestSha: _manifestSha, ltRepoTreeStatus: _errorMesage})
     }
-    console.log("useLtRepoValidation()/useEffect(), refresh=",refresh)
 
     if (authentication && owner && server && languageId) {
-      console.log("useLtRepoValidation() refresh value:",refresh)
       if ( refresh === LT || refresh === ALL ) {
         getReposTrees()
       }
